Added line-by-line comments
