+++ src/commands/list.ts
@@ -3,12 +3,11 @@
 import * as nv from 'node-vault';
 import * as vscode from 'vscode';
 
-export default async function(client: nv.client, path?: string): Promise<string[]> {
+export default async function(client: nv.client, path: string): Promise<string[]> {
+    // List the path
+    const listResponse = await client.list(path);
+    // Parse the list of child paths
+    const children = listResponse.data.keys;
     vscode.window.vault.log(`Listing ${path}`);
-    let children: string[] = [];
-    if (path) {
-        const listResponse = await client.list(path);
-        children = listResponse.data.keys;
-    }
     return children;
 }
