+++ src/commands/login/userpass.ts
@@ -11,11 +11,11 @@ export default async function(client: nv.client): Promise<VaultToken> {
     let token: VaultToken;
     // Prompt the user for the authentication mount point
     const newUserpassMountPoint = await vscode.window.showInputBox({ prompt: 'Enter userpass authentication mount point', value: userpassLoginRequest.mount_point });
-    // If input was collected, cache the input, otherwise cancel
+    // If the username was collected
     if (newUserpassMountPoint) {
         // Prompt the user for the authentication username
         const newUserpassUsername = await vscode.window.showInputBox({ prompt: 'Enter Username', value: userpassLoginRequest.username });
-        // If input was collected, cache the input, otherwise cancel
+        // If the password was collected
         if (newUserpassUsername) {
             // Prompt the user for the authentication password
             const newUserpassPassword = await vscode.window.showInputBox({ password: true, prompt: 'Enter Password', value: userpassLoginRequest.password });
@@ -25,12 +25,14 @@ export default async function(client: nv.client): Promise<VaultToken> {
                 userpassLoginRequest.mount_point = newUserpassMountPoint;
                 userpassLoginRequest.username = newUserpassUsername;
                 userpassLoginRequest.password = newUserpassPassword;
-
-                vscode.window.vault.log('Logging in with username and password', 'person');
+                // Submit a login request
                 const userpassLoginResult = await client.userpassLogin(userpassLoginRequest);
+                // Parse the login response
                 token = { id: userpassLoginResult.auth.client_token, renewable: userpassLoginResult.auth.renewable, ttl: userpassLoginResult.auth.lease_duration };
+                vscode.window.vault.log('Logging in with username and password', 'person');
             }
         }
     }
+    // Return the token (if defined)
     return token;
 }
