--- src/index.d.ts
@@ -1,15 +1,0 @@
-import { VaultToken, VaultWindow } from './model';
-
-declare module 'vscode' {
-    export interface CallableQuickPickItem extends QuickPickItem {
-        callback(...args: any[]) : Promise<VaultToken>;
-    }
-
-    export interface SecretQuickPickItem extends QuickPickItem {
-        secretValue: string;
-    }
-
-    export namespace window {
-        export let vault: VaultWindow;
-    }
-}
