+++ src/model/index.ts
@@ -1,0 +1,5 @@
+'use strict';
+
+export { VaultSession } from './session';
+export { VaultToken } from './token';
+export { VaultWindow } from './window';
