+++ src/model/token.ts
@@ -1,0 +1,7 @@
+'use strict';
+
+export interface VaultToken {
+    id: string;
+    renewable: boolean;
+    ttl: number;
+}
