+++ src/view/treeitem.ts
@@ -35,16 +35,24 @@ export abstract class VaultTreeItem extends vscode.TreeItem {
     }
 
     set children(value: VaultTreeItem[]) {
+        // Set the field to the specified value
         this._children = value;
+        // If the specified value is undefined
         if (value === undefined) {
+            // Update the icon to indicate an error
             this.iconPath = VaultTreeItem.WARNING_ICON;
         }
         else {
+            // Update the icon to use the default icon
             this.iconPath = this._defaultIconPath;
+            // If the specified value is empty
             if (value.length === 0) {
+                // Update the collapsable state
                 this.collapsibleState = vscode.TreeItemCollapsibleState.None;
             }
+            // If the specified value is *not* empty, but was previously empty
             else if (this.collapsibleState === vscode.TreeItemCollapsibleState.None) {
+                // Update the collapsable state
                 this.collapsibleState = vscode.TreeItemCollapsibleState.Collapsed;
             }
         }
